+++ src/routes/staticRoutes.tsx
@@ -1,31 +1,30 @@
 import { BasicLayout } from '@/layouts';
-import React from 'react';
-import { createBrowserRouter } from 'react-router-dom';
+import { lazy } from 'react';
+import { createBrowserRouter, RouteObject } from 'react-router-dom';
 
 // https://github.com/remix-run/react-router/blob/dev/examples/data-router/src/app.tsx
-
-const staticRoutes = createBrowserRouter([
+// https://reactrouter.com/en/main/route/route
+const routes: RouteObject[] = [
   {
-    path: '/',
-    Component: BasicLayout,
+    path: '/', // 路径
+    Component: BasicLayout, //渲染组件
     children: [
       {
         path: '/home',
-        icon: 'home',
-        Component: React.lazy(() => import('@/pages/home')),
+        index: true, // 子路由的默认路由
+        Component: lazy(() => import('@/pages/home')),
       },
       {
         path: '/about',
-        icon: 'home',
-        component: React.lazy(() => import('@/pages/about')),
+        Component: lazy(() => import('@/pages/about')),
       },
     ],
   },
   {
     path: '/404',
-    title: '404',
-    component: React.lazy(() => import('@/pages/error/NoFund404')),
+    Component: lazy(() => import('@/pages/error/NoFund404')),
   },
-]);
+];
+const staticRoutes = createBrowserRouter(routes);
 
-export default staticRoutes;
+export { staticRoutes };